Deleted notes and organized code

+++ Exercise5/ImageGalleryDocumentTableViewController.swift
@@ -11,7 +11,9 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
     
     private lazy var tableModel = GalleryNamesTableModel()
     
+    let imageGalleriesSection = 0
     
+    let deletedImageGalleriesSection = 1
     
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -38,8 +40,8 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
         switch section {
-        case 0: return "Image Galleries"
-        case 1: return "Recently Deleted Image Galleries"
+        case imageGalleriesSection: return "Image Galleries"
+        case deletedImageGalleriesSection: return "Recently Deleted Image Galleries"
         default: return ""
        }
     }
@@ -51,8 +53,8 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // #warning Incomplete implementation, return the number of rows
         switch section {
-        case 0: return tableModel.imageGalleries.count
-        case 1: return tableModel.deletedImageGalleries.count
+        case imageGalleriesSection: return tableModel.imageGalleries.count
+        case deletedImageGalleriesSection: return tableModel.deletedImageGalleries.count
         default: return 0
         }
     }
@@ -64,15 +66,20 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
         
         if let inputCell = cell as? MyTableViewCell {
             switch indexPath.section {
-            case 0: inputCell.textField.text = tableModel.imageGalleries[indexPath.row].galleryName
-            case 1: inputCell.textField.text = tableModel.deletedImageGalleries[indexPath.row].galleryName
+            case imageGalleriesSection: inputCell.textField.text = tableModel.imageGalleries[indexPath.row].galleryName
+            case deletedImageGalleriesSection: inputCell.textField.text = tableModel.deletedImageGalleries[indexPath.row].galleryName
             default: break
             }
             inputCell.resignationHandler = { [weak self, unowned inputCell] in
                 if let text = inputCell.textField.text {
                     switch indexPath.section {
-                    case 0: self?.tableModel.imageGalleries[indexPath.row].galleryName = text
-                    case 1: self?.tableModel.deletedImageGalleries[indexPath.row].galleryName = text
+                    case self?.imageGalleriesSection:
+                        self?.tableModel.imageGalleries[indexPath.row].galleryName = text
+                        self?.tableModel.saveGalleriesInDefaults()
+                        
+                    case self?.deletedImageGalleriesSection:
+                        self?.tableModel.deletedImageGalleries[indexPath.row].galleryName = text
+                        self?.tableModel.saveDeletedGalleriesInDefaults()
                     default: break
                     }
                 }
@@ -80,7 +87,7 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
             }
             
             inputCell.segueHandler = { [weak self, unowned inputCell] in
-                if indexPath.section == 0 {
+                if indexPath.section == self?.imageGalleriesSection {
                 self?.performSegue(withIdentifier: "TableToGallery", sender: inputCell)
                 }
             }
@@ -98,47 +105,24 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
            return 40
     }
 
-    
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
 
-
-    
-//    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-//        if editingStyle == .delete {
-//            // Delete the row from the data source
-//            tableModel.deleteRow(at: indexPath)
-//            self.tableView.deleteRows(at: [indexPath], with: .fade)
-//
-//            tableView.reloadData()
-//        } else if editingStyle == .insert {
-//            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-//        }
-//    }
-    // Override to support editing the table view.
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
             // Delete the row from the data source
             switch indexPath.section {
                 
-            case 0:
+            case imageGalleriesSection:
                 let deletedGallery = tableModel.deleteFromImageGallery(at: indexPath)
                 tableView.deleteRows(at: [indexPath], with: .fade)
                 tableModel.addToDeletedGallery(gallery: deletedGallery)
                 
-            case 1:
+            case deletedImageGalleriesSection:
                 tableModel.deleteFromDeletedImageGallery(at: indexPath)
                 tableView.deleteRows(at: [indexPath], with: .fade)
                 
             default: break
             }
             tableView.reloadData()
-            
         }
     }
     
@@ -148,17 +132,12 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
                 let action = UIContextualAction(style: .normal, title: title,
                     handler: { (action, view, completionHandler) in
                     
-                    switch indexPath.section {
-                        
-                    case 0:break
-                        
-                    case 1:
+                    if indexPath.section == self.deletedImageGalleriesSection {
                         let deletedRow = self.tableModel.deleteFromDeletedImageGallery(at: indexPath)
                         tableView.deleteRows(at: [indexPath], with: .fade)
                         self.tableModel.addToImageGallery(gallery: deletedRow)
-
-                    default: break
                     }
+                    
                     completionHandler(true)
                     tableView.reloadData()
                     
@@ -166,57 +145,11 @@ class ImageGalleryDocumentTableViewController: UITableViewController {
                 let configuration = UISwipeActionsConfiguration(actions: [action])
                 return configuration
         }
-    
-//    // Override to support editing the table view.
-//    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-//        if editingStyle == .delete {
-//            // Delete the row from the data source
-//            tableModel.deleteRow(at: indexPath)
-//            self.tableView.deleteRows(at: [indexPath], with: .fade)
-//
-//            tableView.reloadData()
-//        } else if editingStyle == .insert {
-//            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-//        }
-//    }
-    
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
+
+
 extension Collection {
     func count(where test: (Element) throws -> Bool) rethrows -> Int {
         return try self.filter(test).count
     }
 }
-
-
-//extension ImageGalleryDocumentTableViewController: UITextFieldDelegate {
-//    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
-//        performSegue(withIdentifier: "TableToGallery", sender: textField)
-//        return true
-//    }
-//}
